requirements.txt
fastapi[standard]
uvicorn[standard]
httpx
docker
python-json-logger
pydantic
pydantic-settings
jinja2
pyyaml
watchdog
# ssh-docker-client embedded in src/ directory until published to PyPI
